Committing the full directory with the right scripts
Removed the whole standard folder and moves all the sprites and scripts.
The backup folder in scripts is used only for reference, we can remove
all the files there without affecting anything.

-CharController now controls the character drag and drop the character
on it.
-SmoothFollow now controls the camera, it doesn't feel very smooth
anymore we gotta look into it. Ill doublecheck with the older version,
this may be due to larger camera size.